+++ frontend/src/features/admin/metamodel/index.ts
@@ -2,6 +2,7 @@ export { default as FieldEditorDialog } from "./FieldEditorDialog";
 export { default as StakeholderRolePanel } from "./StakeholderRolePanel";
 export { default as TranslationDialog } from "./TranslationDialog";
 export { default as TypeDetailDrawer } from "./TypeDetailDrawer";
+export { default as RelationTypeValuesDialog } from "./RelationTypeValuesDialog";
 export { default as MetamodelGraph } from "./MetamodelGraph";
 export * from "./constants";
 export * from "./helpers";